--- src/inputs/dropdown/dropdown-btn/type.ts
@@ -1,17 +1,0 @@
-enum BUTTON_STYLE_TYPE {
-    'primary-dark' = 'primary-dark'
-}
-
-export interface DropdownBtnProps {
-    popup: boolean;
-    disabled: boolean;
-    block: boolean;
-    buttonOnly: boolean;
-    buttonIcon?: string;
-    buttonStyleType?: keyof BUTTON_STYLE_TYPE;
-    useCustomStyle: boolean;
-    width: string;
-    height: string;
-    offsetTop: number;
-    showPopup: boolean;
-}
