+++ src/inputs/dropdown/dropdown-menu-btn/PDropdownMenuBtn.vue
@@ -1,31 +1,32 @@
 <template>
-    <div v-click-outside="outsideClick"
-         class="p-dropdown-menu-btn"
-    >
-        <p-dropdown-btn :popup.sync="popup"
-                        :block="block"
-                        :disabled="disabled"
-                        :button-only="buttonOnly"
-                        :button-icon="buttonIcon"
-                        :button-style-type="buttonStyleType"
-                        :position.sync="position"
-                        :use-custom-style="useCustomStyle"
-                        :offset-top.sync="offsetTop"
-                        :width.sync="width"
-                        :height.sync="height"
-                        :show-popup.sync="proxyShowPopup"
-                        @click="$emit('openMenu')"
-        >
-            <template v-for="(_, slot) of buttonSlots" v-slot:[slot]="scope">
-                <slot :name="`button-${slot}`" v-bind="scope" />
-            </template>
-            <template #default>
-                <slot />
-            </template>
-        </p-dropdown-btn>
+    <div v-click-outside="onClickOutside" class="p-dropdown-menu-button">
+        <div ref="dropdownBtn" class="p-dropdown-button" :class="{'button-only': buttonOnly}">
+            <p-button v-if="!buttonOnly"
+                      :disabled="disabled"
+                      :tabindex="-1"
+                      class="menu-button"
+                      :class="{active: popup, hovered: mouseover}"
+                      :style-type="buttonStyleType"
+                      @click="onClick"
+                      @mouseover="onMouseOver"
+                      @mouseout="onMouseOut"
+            >
+                <slot name="default" />
+            </p-button>
+            <p-icon-button :name="buttonIcon || (popup ? 'ic_arrow_top' : 'ic_arrow_bottom')"
+                           color="inherit transparent"
+                           :class="{active: popup, hovered: mouseover}"
+                           :style-type="buttonStyleType"
+                           :disabled="disabled"
+                           :outline="true"
+                           shape="square"
+                           @click="onClick"
+                           @mouseenter="onMouseOver"
+                           @mouseleave="onMouseOut"
+            />
+        </div>
         <p-context-menu v-if="popup"
                         class="menu-ctx"
-                        :class="{block}"
                         :menu="menu"
                         :loading="loading"
                         :auto-height="autoHeight"
@@ -44,15 +45,19 @@
 </template>
 
 <script lang="ts">
+import { reduce } from 'lodash';
 import vClickOutside from 'v-click-outside';
+
 import {
     computed, defineComponent, reactive, toRefs, watch,
 } from '@vue/composition-api';
+
 import PContextMenu from '@/inputs/context-menu/PContextMenu.vue';
-import PDropdownBtn from '@/inputs/dropdown/dropdown-btn/PDropdownBtn.vue';
-import { reduce } from 'lodash';
+import PButton from '@/inputs/buttons/button/PButton.vue';
+import PIconButton from '@/inputs/buttons/icon-button/PIconButton.vue';
 import { ICON_BUTTON_STYLE_TYPE } from '@/inputs/buttons/icon-button/type';
-import addonOptions from '@storybook/addon-console';
+import { DropdownMenuButtonProps } from '@/inputs/dropdown/dropdown-menu-btn/type';
+
 import { makeProxy } from '@/util/composition-helpers';
 
 export default defineComponent({
@@ -60,7 +65,11 @@ export default defineComponent({
     directives: {
         clickOutside: vClickOutside.directive,
     },
-    components: { PDropdownBtn, PContextMenu },
+    components: {
+        PIconButton,
+        PButton,
+        PContextMenu,
+    },
     props: {
         menu: {
             type: [Array, Object],
@@ -71,10 +80,6 @@ export default defineComponent({
             default: false,
         },
         autoHeight: {
-            type: Boolean,
-            default: false,
-        },
-        block: {
             type: Boolean,
             default: false,
         },
@@ -107,8 +112,9 @@ export default defineComponent({
             default: false,
         },
     },
-    setup(props, { emit, slots }) {
+    setup(props: DropdownMenuButtonProps, { emit, slots }) {
         const state = reactive({
+            dropdownBtn: null as HTMLElement | null,
             popup: false,
             width: 0,
             height: 0,
@@ -125,9 +131,38 @@ export default defineComponent({
                 return res;
             }, {})),
             proxyShowPopup: makeProxy('showPopup', props, emit),
+            mouseover: false,
         });
 
-        const outsideClick = (): void => {
+        const setCustomStyle = () => {
+            if (state.dropdownBtn) {
+                const winHeight = window.innerHeight;
+                const rects: any = state.dropdownBtn?.getBoundingClientRect();
+                let position = 'bottom';
+                if (winHeight * 0.9 > rects.top) position = 'top';
+
+                state.position = position;
+                state.offsetTop = rects.top;
+                state.width = rects.width;
+                state.height = rects.height;
+            }
+        };
+
+        const onClick = () => {
+            state.popup = !state.popup;
+            state.proxyShowPopup = false;
+            if (props.useCustomStyle) setCustomStyle();
+        };
+
+        const onMouseOver = () => {
+            if (!props.disabled) state.mouseover = true;
+        };
+
+        const onMouseOut = () => {
+            if (!props.disabled) state.mouseover = false;
+        };
+
+        const onClickOutside = (): void => {
             state.popup = false;
         };
 
@@ -145,19 +180,65 @@ export default defineComponent({
 
         return {
             ...toRefs(state),
-            outsideClick,
+            onClickOutside,
             clickMenuEvent,
+            onClick,
+            onMouseOver,
+            onMouseOut,
+            setCustomStyle,
         };
     },
 });
 </script>
 
 <style lang="postcss">
-.p-dropdown-menu-btn {
+.p-dropdown-menu-button {
     position: relative;
 
-    .menu-ctx.block {
-        width: 100%;
+    .p-dropdown-button {
+        display: inline-flex;
+        min-width: 6.5rem;
+        &.button-only {
+            min-width: unset;
+        }
+
+        .menu-button {
+            @apply border-gray-300 text-gray-900 px-2 justify-start text-left flex-grow font-normal;
+            width: auto;
+            min-width: unset;
+            margin-right: -1px;
+            border-radius: 2px 0 0 2px;
+            &:not(.active).hovered {
+                @apply border-gray-900;
+            }
+            &.active {
+                @apply border-secondary text-secondary;
+            }
+            &.disabled {
+                @apply bg-gray-100 text-gray-300;
+            }
+        }
+
+        .p-icon-button.outline {
+            @apply flex-shrink-0 border-gray-300;
+            &:not(.active).hovered {
+                @apply border-gray-900;
+            }
+            &:not(.disabled):hover {
+                border-color: unset;
+                background-color: unset;
+                color: unset;
+            }
+            &.disabled {
+                @apply text-gray-300;
+            }
+            &.active {
+                @apply border-secondary text-secondary;
+                &:hover {
+                    @apply border-secondary text-secondary;
+                }
+            }
+        }
     }
 }
 </style>
